+++ src/config/counter.php
@@ -20,5 +20,6 @@ return [
         'anonymizeIp' => true, // before hashing the IP, anonymize IP too, default is true
         'ipInEvent' => false, // send IP in event, default is false
         'cacheWidgetsSeconds' => 0, // default is 0. when this is not set or is 0, widgets use default caching system so cached results are used as long as the cached data is valid
+        'headlessToken' => 'ComplexToken', // Token to send from frontend for counting in headless mode
     ],
 ];